web/telegram_bot/requirements.txt
aiosqlite==0.19.0
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dotmap==1.3.30
python-dotenv==1.0.1
ffmpeg-progress-yield==0.7.8
greenlet==3.0.3
idna==3.6
jdatetime==4.1.1
kombu==5.3.5
prompt-toolkit==3.0.43
pyaes==1.6.1
Pyrogram==2.0.106
pyromod==1.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
TgCrypto==1.2.5
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
